Fixed minor bug

+++ pipelines/trapum_fold_and_score_pipeline/webpage_candidate_filter.py
@@ -30,7 +30,10 @@ def make_tarfile(output_path,input_path,name):
 
 
 def remove_dir(dir_name):
-    shutil.rmtree(dir_name)
+    if 'TEMP' in dir_name:
+        shutil.rmtree(dir_name)
+    else:
+        log.error("Directory not deleted. Not a temporary folder!") 
 
 
 def candidate_filter_pipeline(data):
@@ -59,9 +62,9 @@ def candidate_filter_pipeline(data):
        for beam in pointing["beams"]:
            for dp in  (beam["data_products"]):
                xml_list.append(dp["filename"])
-               beam_id_list.append(dp["id"]) 
-               
+               beam_id_list.append(beam["id"]) 
 
+               
    
        # Make temporary folder to keep any temporary outputs
        tmp_dir = '/beeond/PROCESSING/TEMP/%d'%processing_id
@@ -72,16 +75,20 @@ def candidate_filter_pipeline(data):
            pass
 
     
+
          
        # Run the candidate filtering code
        try:
           xml_list2 = ','.join(xml_list)
           subprocess.check_call("candidate_filter.py -i %s -o %s/%d -c /home/psr/software/candidate_filter/candidate_filter/default_config.json --rfi /home/psr/software/candidate_filter/candidate_filter/known_rfi.txt --p_tol %f --dm_tol %f"%(xml_list2,tmp_dir,processing_id,processing_args['p_tol'],processing_args['dm_tol']),shell=True)
+          log.info("Filtered csvs have been written")
        except Exception as error:
           log.error(error)
 
 
        # insert beam ID in good cands to fold csv file for later reference
+
+       log.info("Adding beam id column to folding csv file")
        df = pd.read_csv('%s/%d_good_cands_to_fold.csv'%(tmp_dir,processing_id))
        all_xml_files = df['file'].values
 
@@ -94,14 +101,19 @@ def candidate_filter_pipeline(data):
 
        df['beam_id'] = np.asarray(beam_id_values)
        df.to_csv('%s/%d_good_cands_to_fold_with_beam.csv'%(tmp_dir,processing_id))
+       log.info("New beam id column added to folding csv file")
     
           
        # Tar up the csv files
+       log.info("Tarring up all csvs")
        tar_name = os.path.basename(output_dir)+"_csv_files.tar.gz"
-       make_tarfile(output_dir,tmp_dir,tar_name)    
+       make_tarfile(output_dir,tmp_dir,tar_name)   
+       log.info("Tarred.")
+        
 
        # Remove contents in temporary directory
        remove_dir(tmp_dir)  
+       log.info("Removed temporary files")
 
        # Add tar file to dataproduct
        dp = dict(
